Import initial version of tube client
Implement a bidirectional authenticated peer-to-peer pipe modelled after
Magic Wormhole. Our security story is similar, with 20-bit base32 keys
comparable in entropy to Magic Wormhole's 16-bit English word pairs, a
forward-secure authenticated key exchange, and explicit verification
after pairing, before any data is sent.

The code comprises a ticket number plus 4-character base32 key, and the
initiator displays it in the form '1-4mdf'. When the responder prompts
for the code, it ignores case and the two halves can be separated by
whitespace and/or any punctuation characters.

After hashing the base32 key into the secret state, we use map-to-point
SPEKE as described in the Pocketcrypt README for authenticated key
exchange. This is now standardised as CPace and thoroughly analysed in
https://eprint.iacr.org/2021/114.pdf.

The shared secret state is explicitly authenticated, forking it for the
transmit and receive streams by absorbing our ephemeral identity or our
peer's identity respectively, then squeezing and verifying tags in both
directions. The broken symmetry and sponge non-linearity thwarts
attacks like https://eprint.iacr.org/2014/585.pdf although parallel key
re-use is not possible in this application anyway.

The shared secret state is also used to generate a verifier. Confirming
this out-of-band with our peer will detect man-in-the-middle attacks
even if the original code was guessed or leaked.

Once the tunnel is established, transmission and reception can proceed
independently. Spawn worker threads so these potentially cpu-bound
workloads can run concurrently. For each chunk of up to 65518 bytes
read from stdin, we send its 16-bit little-endian length, followed by
the duplex-encrypted data payload. After padding the resulting state, a
16-byte tag is squeezed and appended in the usual way. An authenticated
block whose payload length is zero indicates the end of the stream.